test(autoware_traffic_light_visualization): retire characterization test and rename core unit test (#12612)
* test(autoware_traffic_light_visualization): retire characterization test for TrafficLightMapVisualizerNode

The characterization test introduced in #12487 has served its purpose
as the safety net during the refactor series (#12511 / #12553 / #12591
/ #12596). Its scope is now fully covered by the finer-grained layers:

- extract_bulbs unit tests: lanelet -> Bulb conversion
- TrafficLightVisualizer core unit tests: bulb -> marker generation
- Node smoke tests: pub/sub plumbing and map-not-received guard

Remove the characterization test and its CMakeLists.txt entry. The
remaining test executables run an order of magnitude faster, with no
loss of behavioural coverage.

* test(autoware_traffic_light_visualization): rename TrafficLightVisualizer unit test to take the package's main test slot

With the characterization test retired, the unit test file for
TrafficLightVisualizer is the most natural fit for the package's main
test name. Rename it from test_traffic_light_map_visualizer_core.cpp
to test_traffic_light_map_visualizer.cpp and update the matching
ament_add_gtest target name in CMakeLists.txt accordingly.

No content changes; this is a pure rename + CMake target rename.
